Make the NativeLoginManager instance nullable instead of using lateinit.  Rename the package containing the NativeLoginManager interface.
